pyproject.toml
[project]
name = "dict-grpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio-tools>=1.69.0",
    "grpcio>=1.69.0",
]
